refactor Lolsearch Service -> Infrastructure의 Adapter로 변경
